LayoutTests/imported/w3c/web-platform-tests/eventsource/dedicated-worker/eventsource-prototype.htm

<!-- 
try {
  EventSource.prototype.ReturnTrue = function() { return true }
  var source = new EventSource("../resources/message.py")
  postMessage([true, source.ReturnTrue(), 'EventSource' in self])
  source.close()
} catch(e) {
  postMessage([false, String(e)])
}
/* -->
<!DoCtYpE hTMl>
<html>
  <heAd>
    <title>dedicated worker - EventSource: prototype et al</title>
    <scrIpt src="/resources/testharness.js"></script>
    <scriPt src="/resources/testharnessreport.js"></script>
  </head>
  <boDy>
    <diV iD="log"></div>
    <sCript>
      var test = async_test();
      test.step(function() {
        var worker = new Worker('#')
        worker.onmessage = function(e) {
          test.step(function() {
            assert_true(e.data[0], e.data[1])
            assert_true(e.data[1], 'source.ReturnTrue()')
            assert_true(e.data[2], "'EventSource' in self")
          })
          test.done()
        }
      })
    </script>
  </body>
</html>
<!-- */ // -->
